Refactor `Spoiler` into a separate module
