+++ WebRandomizer/ClientApp/src/spoiler/Spoiler.jsx
